Proyecto final Api

+++ app/Http/Controllers/Admin/Doctor/DoctorsController.php
@@ -3,8 +3,10 @@
 namespace App\Http\Controllers\Admin\Doctor;
 
 use App\Http\Controllers\Controller;
+use App\Http\Resources\Appointment\AppointmentCollection;
 use App\Http\Resources\User\UserCollection;
 use App\Http\Resources\User\UserResource;
+use App\Models\Appointment\Appointment;
 use App\Models\Doctor\DoctorScheduleDay;
 use App\Models\Doctor\DoctorScheduleHour;
 use App\Models\Doctor\DoctorScheduleJoinHour;
@@ -13,6 +15,7 @@ use App\Models\User;
 use Carbon\Carbon;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\Redis;
 use Illuminate\Support\Facades\Storage;
 use Spatie\Permission\Models\Role;
 
@@ -23,6 +26,7 @@ class DoctorsController extends Controller
      */
     public function index(Request $request)
     {
+        $this->authorize('viewAnyDoctor', Doctor::class);
         $search = $request->search;
 
         $users = User::where(DB::raw("CONCAT(users.name,' ',IFNULL(users.surname,''),' ',users.email)"), "like", "%" . $search . "%")
@@ -40,6 +44,86 @@ class DoctorsController extends Controller
         ]);
     }
 
+    public function profile($id)
+    {
+        $this->authorize('profileDoctor', Doctor::class);
+        $cachedRecord = Redis::get('profile_doctor_#' . $id);
+        $data_doctor = [];
+        if (isset($cachedRecord)) {
+            $data_doctor = json_decode($cachedRecord, FALSE);
+        } else {
+            $user = User::findOrFail($id);
+
+            $num_appointment = Appointment::where("doctor_id", $id)->count();
+            $money_of_appointments = Appointment::where("doctor_id", $id)->sum("amount");
+            $num_appointment_pendings = Appointment::where("doctor_id", $id)->where("status", 1)->count();
+
+            $appointment_pendings = Appointment::where("doctor_id", $id)->where("status", 1)->orderBy("date_appointment", "asc")->get();
+            $appointments = Appointment::where("doctor_id", $id)->orderBy("date_appointment", "asc")->get();
+
+            $data_doctor = [
+                "num_appointment" => $num_appointment,
+                "money_of_appointments" => $money_of_appointments,
+                "num_appointment_pendings" => $num_appointment_pendings,
+                "doctor" => UserResource::make($user),
+                "appointment_pendings" => AppointmentCollection::make($appointment_pendings),
+                "appointments" => $appointments->map(function ($appointment) {
+                    date_default_timezone_set('America/Argentina/Buenos_Aires');
+                    Carbon::setLocale('es'); // Establecer el idioma a español
+                    $months = [
+                        'Jan' => 'Ene',
+                        'Feb' => 'Feb',
+                        'Mar' => 'Mar',
+                        'Apr' => 'Abr',
+                        'May' => 'May',
+                        'Jun' => 'Jun',
+                        'Jul' => 'Jul',
+                        'Aug' => 'Ago',
+                        'Sep' => 'Sep',
+                        'Oct' => 'Oct',
+                        'Nov' => 'Nov',
+                        'Dec' => 'Dic'
+                    ];
+
+                    $date = Carbon::parse($appointment->date_appointment);
+                    $date_format = $date->format('d M Y');
+
+                    // Reemplazar el mes en inglés con la abreviatura en español
+                    $date_appointment_format = strtr($date_format, $months);
+                    return [
+                        "id" => $appointment->id,
+                        "patient" => [
+                            "id" => $appointment->patient->id,
+                            "full_name" => $appointment->patient->name . ' ' . $appointment->patient->surname,
+                            "avatar" => $appointment->patient->avatar ? env("APP_URL") . "storage/" . $appointment->patient->avatar : 'https://cdn-icons-png.flaticon.com/512/1430/1430453.png',
+                        ],
+                        "doctor" => [
+                            "id" => $appointment->doctor->id,
+                            "full_name" => $appointment->doctor->name . ' ' . $appointment->doctor->surname,
+                            "avatar" => $appointment->doctor->avatar ? env("APP_URL") . "storage/" . $appointment->doctor->avatar : NULL,
+                        ],
+                        "date_appointment" => $appointment->date_appointment,
+                        "date_appointment_format" => $date_appointment_format,
+                        "format_hour_start" => Carbon::parse(date("Y-m-d") . ' ' . $appointment->doctor_schedule_join_hour->doctor_schedule_hour->hour_start)->format("h:i A"),
+                        "format_hour_end" => Carbon::parse(date("Y-m-d") . ' ' . $appointment->doctor_schedule_join_hour->doctor_schedule_hour->hour_end)->format("h:i A"),
+                        "appointment_attention" => $appointment->attention ? [
+                            "id" => $appointment->attention->id,
+                            "description" => $appointment->attention->description,
+                            "receta_medica" => $appointment->attention->receta_medica ? json_decode($appointment->attention->receta_medica) : [],
+                            "created_at" => $appointment->attention->created_at->format("Y-m-d h:i A"),
+                        ] : NULL,
+                        "amount" => $appointment->amount,
+                        "status_pay" => $appointment->status_pay,
+                        "status" => $appointment->status,
+                    ];
+                }),
+            ];
+            Redis::set('profile_doctor_#' . $id, json_encode($data_doctor), 'EX', 3600);
+        }
+
+
+        return response()->json($data_doctor);
+    }
 
     public function config()
     {
@@ -74,11 +158,14 @@ class DoctorsController extends Controller
         ]);
     }
 
+
     /**
      * Store a newly created resource in storage.
      */
     public function store(Request $request)
     {
+        $this->authorize('createDoctor', Doctor::class);
+
         $schedule_hours = json_decode($request->schedule_hours, 1);
 
         $users_is_valid = User::where("email", $request->email)->first();
@@ -136,6 +223,8 @@ class DoctorsController extends Controller
      */
     public function show(string $id)
     {
+        $this->authorize('viewDoctor', Doctor::class);
+
         $user = User::findOrFail($id);
 
         return response()->json([
@@ -149,6 +238,8 @@ class DoctorsController extends Controller
      */
     public function update(Request $request, string $id)
     {
+        $this->authorize('updateDoctor', Doctor::class);
+
         $schedule_hours = json_decode($request->schedule_hours, 1);
 
         $users_is_valid = User::where("id", "<>", $id)->where("email", $request->email)->first();
@@ -179,6 +270,10 @@ class DoctorsController extends Controller
         $request->request->add(["birth_date" => Carbon::parse($date_clean)->format("Y-m-d h:i:s")]);
 
         // $request->request->add(["birth_date" => Carbon::parse($request->birth_date, 'GMT')->format("Y-m-d h:i:s")]);
+        $cachedRecord = Redis::get('profile_doctor_#' . $id);
+        if (isset($cachedRecord)) {
+            Redis::del('profile_doctor_#' . $id);
+        }
         $user->update($request->all());
 
         if ($request->role_id != $user->roles()->first()->id) {
@@ -315,8 +410,14 @@ class DoctorsController extends Controller
      */
     public function destroy(string $id)
     {
+        $this->authorize('deleteDoctor', Doctor::class);
+
         $user = User::findOrFail($id);
         $user->delete();
+        $cachedRecord = Redis::get('profile_doctor_#' . $id);
+        if (isset($cachedRecord)) {
+            Redis::del('profile_doctor_#' . $id);
+        }
         return response()->json([
             "message" => 200
         ]);
