+++ app/Http/Controllers/Admin/Rol/RolesController.php
@@ -2,9 +2,9 @@
 
 namespace App\Http\Controllers\Admin\Rol;
 
-use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 use Spatie\Permission\Models\Role;
+use App\Http\Controllers\Controller;
 
 class RolesController extends Controller
 {
@@ -14,6 +14,9 @@ class RolesController extends Controller
     public function index(Request $request)
     {
         // QUE EL FILTRO POR NOMBRE DE ROL
+        if (!auth('api')->user()->can('list_rol')) {
+            return response()->json(["message" => "EL USUARIO NO ESTA AUTORIZADO"], 403);
+        }
         $name = $request->search;
 
         $roles = Role::where("name", "like", "%" . $name . "%")->orderBy("id", "desc")->get();
@@ -36,12 +39,15 @@ class RolesController extends Controller
      */
     public function store(Request $request)
     {
+        if (!auth('api')->user()->can('register_rol')) {
+            return response()->json(["message" => "EL USUARIO NO ESTA AUTORIZADO"], 403);
+        }
         $is_role = Role::where("name", $request->name)->first();
 
         if ($is_role) {
             return response()->json([
                 "message" => 403,
-                "message_text" => "El nombre del rol ya existe."
+                "message_text" => "EL NOMBRE DEL ROL YA EXISTE"
             ]);
         }
 
@@ -63,6 +69,9 @@ class RolesController extends Controller
      */
     public function show(string $id)
     {
+        if (!auth('api')->user()->can('edit_rol')) {
+            return response()->json(["message" => "EL USUARIO NO ESTA AUTORIZADO"], 403);
+        }
         $role = Role::findOrFail($id);
         return response()->json([
             "id" => $role->id,
@@ -78,12 +87,15 @@ class RolesController extends Controller
      */
     public function update(Request $request, string $id)
     {
+        if (!auth('api')->user()->can('edit_rol')) {
+            return response()->json(["message" => "EL USUARIO NO ESTA AUTORIZADO"], 403);
+        }
         $is_role = Role::where("id", "<>", $id)->where("name", $request->name)->first();
 
         if ($is_role) {
             return response()->json([
                 "message" => 403,
-                "message_text" => "El nombre del rol ya existe."
+                "message_text" => "EL NOMBRE DEL ROL YA EXISTE"
             ]);
         }
 
@@ -102,11 +114,14 @@ class RolesController extends Controller
      */
     public function destroy(string $id)
     {
+        if (!auth('api')->user()->can('delete_rol')) {
+            return response()->json(["message" => "EL USUARIO NO ESTA AUTORIZADO"], 403);
+        }
         $role = Role::findOrFail($id);
         if ($role->users->count() > 0) {
             return response()->json([
                 "message" => 403,
-                "message_text" => "El rol seleccionado no se puede eliminar por motivos que ya tiene usuarios relacionados."
+                "message_text" => "EL ROL SELECCIONADO NO SE PUEDE ELIMINAR POR MOTIVOS QUE YA TIENE USUARIOS RELACIONADOS"
             ]);
         }
         $role->delete();
