+++ app/Http/Controllers/Admin/Staff/StaffsController.php
@@ -19,6 +19,8 @@ class StaffsController extends Controller
      */
     public function index(Request $request)
     {
+        $this->authorize('viewAny', User::class);
+
         $search = $request->search;
 
         $users = User::where(DB::raw("CONCAT(users.name,' ',IFNULL(users.surname,''),' ',users.email)"), "like", "%" . $search . "%")
@@ -49,6 +51,7 @@ class StaffsController extends Controller
      */
     public function store(Request $request)
     {
+        $this->authorize('create', User::class);
 
         $users_is_valid = User::where("email", $request->email)->first();
 
@@ -87,6 +90,8 @@ class StaffsController extends Controller
      */
     public function show(string $id)
     {
+        $this->authorize('view', User::class);
+
         $user = User::findOrFail($id);
 
         return response()->json([
@@ -99,6 +104,8 @@ class StaffsController extends Controller
      */
     public function update(Request $request, string $id)
     {
+        $this->authorize('update', User::class);
+
         $users_is_valid = User::where("id", "<>", $id)->where("email", $request->email)->first();
 
         if ($users_is_valid) {
@@ -122,14 +129,15 @@ class StaffsController extends Controller
             $request->request->add(["password" => bcrypt($request->password)]);
         }
 
-        $date_clean = preg_replace('/\(.*\)|[A-Z]{3}-\d{4}/', '', $request->birth_date);
-
-        $request->request->add(["birth_date" => Carbon::parse($date_clean)->format("Y-m-d h:i:s")]);
+        if ($request->birth_date) {
+            $date_clean = preg_replace('/\(.*\)|[A-Z]{3}-\d{4}/', '', $request->birth_date);
+            $request->request->add(["birth_date" => Carbon::parse($date_clean)->format("Y-m-d h:i:s")]);
+        }
 
         // $request->request->add(["birth_date" => Carbon::parse($request->birth_date, 'GMT')->format("Y-m-d h:i:s")]);
         $user->update($request->all());
 
-        if ($request->role_id != $user->roles()->first()->id) {
+        if ($request->role_id && $request->role_id != $user->roles()->first()->id) {
             $role_old = Role::findOrFail($user->roles()->first()->id);
             $user->removeRole($role_old);
 
@@ -146,6 +154,8 @@ class StaffsController extends Controller
      */
     public function destroy(string $id)
     {
+        $this->authorize('delete', User::class);
+
         $user = User::findOrFail($id);
         if ($user->avatar) {
             Storage::delete($user->avatar);
