+++ app/Http/Controllers/Patient/PatientController.php
@@ -3,10 +3,13 @@
 namespace App\Http\Controllers\Patient;
 
 use App\Http\Controllers\Controller;
+use App\Http\Resources\Appointment\AppointmentCollection;
 use App\Http\Resources\Patient\PatientCollection;
 use App\Http\Resources\Patient\PatientResource;
+use App\Models\Appointment\Appointment;
 use App\Models\Patient\Patient;
 use App\Models\Patient\PatientPerson;
+use Illuminate\Support\Facades\Redis;
 use Carbon\Carbon;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
@@ -19,11 +22,13 @@ class PatientController extends Controller
      */
     public function index(Request $request)
     {
+        // $this->authorize('viewAny', Patient::class);
+
         $search = $request->search;
 
         $patients = Patient::where(DB::raw("CONCAT(patients.name,' ',IFNULL(patients.surname,''),' ',patients.email)"), "like", "%" . $search . "%")
             ->orderBy("id", "desc")
-            ->paginate(8);
+            ->paginate(20);
 
         return response()->json([
             "total" => $patients->total(),
@@ -31,11 +36,97 @@ class PatientController extends Controller
         ]);
     }
 
+    public function profile($id)
+    {
+        $this->authorize('profile', Patient::class);
+
+        $cachedRecord = Redis::get('profile_patient_#' . $id);
+        $data_patient = [];
+        if (isset($cachedRecord)) {
+            $data_patient = json_decode($cachedRecord, FALSE);
+        } else {
+            $patient = Patient::findOrFail($id);
+
+            $num_appointment = Appointment::where("patient_id", $id)->count();
+            $money_of_appointments = Appointment::where("patient_id", $id)->sum("amount");
+            $num_appointment_pendings = Appointment::where("patient_id", $id)->where("status", 1)->count();
+
+            $appointment_pendings = Appointment::where("patient_id", $id)->where("status", 1)->orderBy("date_appointment", "asc")->get();
+            $appointments = Appointment::where("patient_id", $id)
+                ->orderBy("date_appointment", "asc")
+                ->get();
+
+            $data_patient = [
+                "num_appointment" => $num_appointment,
+                "money_of_appointments" => $money_of_appointments,
+                "num_appointment_pendings" => $num_appointment_pendings,
+                "patient" => PatientResource::make($patient),
+                "appointment_pendings" => AppointmentCollection::make($appointment_pendings),
+                "appointments" => $appointments->map(function ($appointment) {
+                    date_default_timezone_set('America/Argentina/Buenos_Aires');
+                    Carbon::setLocale('es'); // Establecer el idioma a español
+                    $months = [
+                        'Jan' => 'Ene',
+                        'Feb' => 'Feb',
+                        'Mar' => 'Mar',
+                        'Apr' => 'Abr',
+                        'May' => 'May',
+                        'Jun' => 'Jun',
+                        'Jul' => 'Jul',
+                        'Aug' => 'Ago',
+                        'Sep' => 'Sep',
+                        'Oct' => 'Oct',
+                        'Nov' => 'Nov',
+                        'Dec' => 'Dic'
+                    ];
+
+                    $date = Carbon::parse($appointment->date_appointment);
+                    $date_format = $date->format('d M Y');
+
+                    // Reemplazar el mes en inglés con la abreviatura en español
+                    $date_appointment_format = strtr($date_format, $months);
+                    return [
+                        "id" => $appointment->id,
+                        "patient" => [
+                            "id" => $appointment->patient->id,
+                            "full_name" => $appointment->patient->name . ' ' . $appointment->patient->surname,
+                            "avatar" => $appointment->patient->avatar ? env("APP_URL") . "storage/" . $appointment->patient->avatar : 'assets/img/user.jpg',
+                        ],
+                        "doctor" => [
+                            "id" => $appointment->doctor->id,
+                            "full_name" => $appointment->doctor->name . ' ' . $appointment->doctor->surname,
+                            "avatar" => $appointment->doctor->avatar ? env("APP_URL") . "storage/" . $appointment->doctor->avatar : 'assets/img/user.jpg',
+                        ],
+                        "date_appointment" => $appointment->date_appointment,
+                        "date_appointment_format" => $date_appointment_format,
+                        "format_hour_start" => Carbon::parse(date("Y-m-d") . ' ' . $appointment->doctor_schedule_join_hour->doctor_schedule_hour->hour_start)->format("h:i A"),
+                        "format_hour_end" => Carbon::parse(date("Y-m-d") . ' ' . $appointment->doctor_schedule_join_hour->doctor_schedule_hour->hour_end)->format("h:i A"),
+                        "appointment_attention" => $appointment->attention ? [
+                            "id" => $appointment->attention->id,
+                            "description" => $appointment->attention->description,
+                            "receta_medica" => $appointment->attention->receta_medica ? json_decode($appointment->attention->receta_medica) : [],
+                            "created_at" => $appointment->attention->created_at->format("Y-m-d h:i A"),
+                        ] : NULL,
+                        "amount" => $appointment->amount,
+                        "status_pay" => $appointment->status_pay,
+                        "status" => $appointment->status,
+                    ];
+                }),
+            ];
+            Redis::set('profile_patient_#' . $id, json_encode($data_patient), 'EX', 3600);
+        }
+
+        return response()->json($data_patient);
+    }
+
+
     /**
      * Store a newly created resource in storage.
      */
     public function store(Request $request)
     {
+        $this->authorize('create', Patient::class);
+
         $patient_is_valid = Patient::where("n_document", $request->n_document)->first();
 
         if ($patient_is_valid) {
@@ -73,6 +164,8 @@ class PatientController extends Controller
      */
     public function show(string $id)
     {
+        $this->authorize('view', Patient::class);
+
         $patient = Patient::findOrFail($id);
 
         return response()->json([
@@ -85,6 +178,8 @@ class PatientController extends Controller
      */
     public function update(Request $request, string $id)
     {
+        $this->authorize('update', Patient::class);
+
         $patient_is_valid = Patient::where("id", "<>", $id)->where("n_document", $request->n_document)->first();
 
         if ($patient_is_valid) {
@@ -110,6 +205,11 @@ class PatientController extends Controller
             $request->request->add(["birth_date" => Carbon::parse($date_clean)->format("Y-m-d h:i:s")]);
         }
 
+        $cachedRecord = Redis::get('profile_patient_#' . $id);
+        if (isset($cachedRecord)) {
+            Redis::del('profile_patient_#' . $id);
+        }
+
         // $request->request->add(["birth_date" => Carbon::parse($request->birth_date, 'GMT')->format("Y-m-d h:i:s")]);
         $patient->update($request->all());
 
@@ -126,9 +226,15 @@ class PatientController extends Controller
      */
     public function destroy(string $id)
     {
+        $this->authorize('delete', Patient::class);
+
         $patient = Patient::findOrFail($id);
         if ($patient->avatar) {
             Storage::delete($patient->avatar);
+        }
+        $cachedRecord = Redis::get('profile_patient_#' . $id);
+        if (isset($cachedRecord)) {
+            Redis::del('profile_patient_#' . $id);
         }
         $patient->delete();
         return response()->json([
